[be] fix: feedbackboard  1. 피드백 수정, 삭제시 멤버 아이디 확인절차 추가  2. 피드백 카테고리 목록 조회 결과와 pageInfo가 맞게 출력되도록 수정  3. 주석 추가

+++ server/src/main/java/com/CreatorConnect/server/feedbackboard/service/FeedbackBoardService.java
@@ -16,10 +16,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -32,7 +35,7 @@ public class FeedbackBoardService {
     private final FeedbackCategoryRepository feedbackCategoryRepository;
     private final FeedbackCategoryService feedbackCategoryService;
 
-    //post
+    //등록
     public FeedbackBoardResponseDto.Post createFeedback(FeedbackBoardDto.Post postDto){
         // Dto-Entity 변환
         FeedbackBoard feedbackBoard = mapper.feedbackBoardPostDtoToFeedbackBoard(postDto);
@@ -44,17 +47,28 @@ public class FeedbackBoardService {
         //저장
         FeedbackBoard savedfeedbackBoard = feedbackBoardRepository.save(feedbackBoard);
 
-        // Entity-Dto 변환
+        // Entity-Dto 변환 후 리턴
         FeedbackBoardResponseDto.Post responseDto = mapper.feedbackBoardToFeedbackBoardPostResponse(savedfeedbackBoard);
         return responseDto;
     }
 
-    //patch
+    //수정
     public FeedbackBoardResponseDto.Patch updateFeedback(Long feedbackBoardId, FeedbackBoardDto.Patch patchDto){
+
+        // Dto-Entity 변환
         FeedbackBoard feedbackBoard = mapper.feedbackBoardPatchDtoToFeedbackBoard(patchDto);
         feedbackBoard.setFeedbackBoardId(feedbackBoardId);
+
+        // Dto의 Id값으로 데이터베이스에 있는 Entity 찾기
         FeedbackBoard foundFeedbackBoard = findVerifiedFeedbackBoard(feedbackBoard.getFeedbackBoardId());
 
+        // 글 작성한 멤버가 현재 로그인한 멤버와 같은지 확인
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(!Objects.equals(foundFeedbackBoard.getMember().getEmail(), authentication.getName())){
+            throw new BusinessLogicException(ExceptionCode.INVALID_MEMBER);
+        }
+
+        // 찾은 Entity의 값 변경
         Optional.ofNullable(feedbackBoard.getTitle())
                 .ifPresent(foundFeedbackBoard::setTitle);
         Optional.ofNullable(feedbackBoard.getLink())
@@ -65,42 +79,66 @@ public class FeedbackBoardService {
         foundFeedbackBoard.setCategory(category.orElseThrow(() -> new BusinessLogicException(ExceptionCode.CATEGORY_NOT_FOUND)));
         Optional<FeedbackCategory> feedbackCategory = feedbackCategoryRepository.findByFeedbackCategoryName(patchDto.getFeedbackCategoryName());
         foundFeedbackBoard.setFeedbackCategory(feedbackCategory.orElseThrow(() -> new BusinessLogicException(ExceptionCode.FEEDBACK_CATEGORY_NOT_FOUND)));
-        // Todo FeedbackBoard-tag연결(2)
-//        Optional.ofNullable(feedbackBoard.getTag())
-//                .ifPresent(foundFeedbackBoard::setTag);
+        Optional.ofNullable(feedbackBoard.getTag())
+                .ifPresent(foundFeedbackBoard::setTag);
 
+        // 저장
         FeedbackBoard updatedFeedbackBoard = feedbackBoardRepository.save(foundFeedbackBoard);
+
+        // Entity-Dto 변환 후 리턴
         FeedbackBoardResponseDto.Patch responseDto = mapper.feedbackBoardToFeedbackBoardPatchResponse(updatedFeedbackBoard);
         return responseDto;
     }
 
-    // 하나 조회
+    // 개별 조회
     public FeedbackBoardResponseDto.Details responseFeedback(Long FeedbackBoardId){
+        // 클라이언트에서 보낸 ID값으로 데이터베이스에 있는 Entity 찾아서 리턴
         FeedbackBoard foundFeedbackBoard = findVerifiedFeedbackBoard(FeedbackBoardId);
         return mapper.feedbackBoardToFeedbackBoardDetailsResponse(foundFeedbackBoard);
     }
 
     //목록 조회
     public FeedbackBoardResponseDto.Multi<FeedbackBoardResponseDto.Details> responseFeedbacks(int page, int size){
-        PageRequest pageRequest = PageRequest.of(page - 1, size, Sort.by("feedbackBoardId").descending());
-        Page<FeedbackBoard> feedbackBoardsPage = feedbackBoardRepository.findAll(pageRequest);
+        //Page 생성 - 피드백 보드Id 기준 내림차순 정렬 적용
+        Page<FeedbackBoard> feedbackBoardsPage = feedbackBoardRepository.findAll(PageRequest.of(page - 1, size, Sort.by("feedbackBoardId").descending()));
+
+        // 피드백 리스트 가져오기
         List<FeedbackBoardResponseDto.Details> responses = mapper.feedbackBoardsToFeedbackBoardDetailsResponses(feedbackBoardsPage.getContent());
+
+        // pageInfo 가져오기
         FeedbackBoardResponseDto.PageInfo pageInfo = new FeedbackBoardResponseDto.PageInfo(feedbackBoardsPage.getNumber() + 1, feedbackBoardsPage.getSize(), feedbackBoardsPage.getTotalElements(), feedbackBoardsPage.getTotalPages());
+
+        // 리턴
         return new FeedbackBoardResponseDto.Multi<>(responses, pageInfo);
     }
 
     //피드백 카테고리로 목록 조회
     public FeedbackBoardResponseDto.Multi<FeedbackBoardResponseDto.Details> responseFeedbacksByCategory(Long feedbackCategoryId, int page, int size){
-        PageRequest pageRequest = PageRequest.of(page - 1, size, Sort.by("feedbackBoardId").descending());
-        Page<FeedbackBoard> feedbackBoardsPage = feedbackBoardRepository.findAll(pageRequest);
-        List<FeedbackBoardResponseDto.Details> responses = findFeedbacksByCategoryId(feedbackCategoryId);
+        //page생성 - 피드백 카테고리 ID로 검색 후 피드백 보드Id 기준 내림차순 정렬 적용
+        Page<FeedbackBoard> feedbackBoardsPage = feedbackBoardRepository.findFeedbackBoardsByFeedbackCategoryId(feedbackCategoryId, PageRequest.of(page - 1, size, Sort.by("feedbackBoardId").descending()));
+
+        // 피드백 리스트 가져오기
+        List<FeedbackBoardResponseDto.Details> responses = mapper.feedbackBoardsToFeedbackBoardDetailsResponses(feedbackBoardsPage.getContent());
+
+        // pageInfo 가져오기
         FeedbackBoardResponseDto.PageInfo pageInfo = new FeedbackBoardResponseDto.PageInfo(feedbackBoardsPage.getNumber() + 1, feedbackBoardsPage.getSize(), feedbackBoardsPage.getTotalElements(), feedbackBoardsPage.getTotalPages());
+
+        //리턴
         return new FeedbackBoardResponseDto.Multi<>(responses, pageInfo);
     }
 
     //삭제
     public void deleteFeedback(Long feedbackBoardId) {
+        // 피드백 ID로 피드백 찾기
         FeedbackBoard feedbackBoard = findVerifiedFeedbackBoard(feedbackBoardId);
+
+        // 글 작성한 멤버가 현재 로그인한 멤버와 같은지 확인
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(!Objects.equals(feedbackBoard.getMember().getEmail(), authentication.getName())){
+            throw new BusinessLogicException(ExceptionCode.INVALID_MEMBER);
+        }
+
+        // 삭제
         feedbackBoardRepository.delete(feedbackBoard);
     }
 
@@ -111,9 +149,4 @@ public class FeedbackBoardService {
         return FeedbackBoard.orElseThrow(() -> new BusinessLogicException(ExceptionCode.FEEDBACK_NOT_FOUND));
     }
 
-    //피드백 카테고리 아이디로 피드백 목록 찾는 메서드
-    public List<FeedbackBoardResponseDto.Details> findFeedbacksByCategoryId(Long feedbackCategoryId) {
-        FeedbackCategory feedbackCategory = feedbackCategoryService.findVerifiedFeedbackCategory(feedbackCategoryId);
-        return mapper.feedbackBoardsToFeedbackBoardDetailsResponses(feedbackBoardRepository.findByFeedbackCategory(feedbackCategory));
-    }
 }
